Fix/pss events (#659)
* fix pss view source and click though events
