+++ components/PrimarySourceSetsComponents/Source/components/ContentAndMetadata/GaPssWrapper.js
@@ -3,32 +3,21 @@ import ReactGA from "react-ga";
 import Router from "next/router";
 import { gaTrackingId } from "constants/env";
 import {
-  bindLinkEvent,
-  initGa,
   getFullPath,
   joinIfArray,
   getItemId,
   getPartner,
-  trackGaEvent
+  trackGaEvent,
+  getContributor,
+  getTitle,
+  initGa
 } from "utilFunctions";
-
-const getContributor = source =>
-  source.mainEntity[0]["provider"].filter(
-    ref => ref["disambiguationDescription"] == "contributing institution"
-  )[0]["name"];
-
-const getTitle = source => source.mainEntity[0]["name"];
 
 export default WrappedComponent =>
   class GaPssWrapper extends React.Component {
     constructor(props) {
       super(props);
       this.trackSourceView = this.trackSourceView.bind(this);
-      const source = this.props.source;
-      this.itemId = getItemId(source);
-      this.title = joinIfArray(getTitle(source));
-      this.partner = joinIfArray(getPartner(source));
-      this.contributor = joinIfArray(getContributor(source));
     }
 
     // Using componentDidMount enables access to the window, which is necessary
@@ -37,7 +26,6 @@ export default WrappedComponent =>
       initGa();
       this.trackSourceView();
       Router.router.events.on("routeChangeComplete", this.trackSourceView);
-      this.bindClickThroughEvent();
     }
 
     // Cleanup, prevents multiple pageviews being counted for a single route.
@@ -47,33 +35,20 @@ export default WrappedComponent =>
 
     trackSourceView() {
       const fullPath = getFullPath();
+      const source = this.props.source;
 
       if (fullPath !== this.lastTrackedPath) {
         const gaEvent = {
           type: "View Primary Source",
-          itemId: this.itemId,
-          title: this.title,
-          partner: this.partner,
-          contributor: this.contributor
+          itemId: getItemId(source),
+          title: joinIfArray(getTitle(source)),
+          partner: joinIfArray(getPartner(source)),
+          contributor: joinIfArray(getContributor(source))
         };
 
         trackGaEvent(gaEvent);
         this.lastTrackedPath = fullPath;
       }
-    }
-
-    bindClickThroughEvent() {
-      const links = document.getElementsByClassName("clickThrough");
-
-      const gaEvent = {
-        type: "Click Through",
-        itemId: this.itemId,
-        title: this.title,
-        partner: this.partner,
-        contributor: this.contributor
-      };
-
-      bindLinkEvent(gaEvent, links);
     }
 
     render() {
