+++ components/PrimarySourceSetsComponents/Source/components/ContentAndMetadata/index.js
@@ -14,7 +14,10 @@ import {
   joinIfArray,
   getItemId,
   getPartner,
-  showMoreDescription
+  showMoreDescription,
+  getTitle,
+  getContributor,
+  trackGaEvent
 } from "utilFunctions";
 
 import { classNames, stylesheet } from "./ContentAndMetadata.css";
@@ -55,6 +58,32 @@ const getViewerComponent = (fileFormat, type, pathToFile) => {
 
 const getDomainFromThumbnail = thumbnailUrl =>
   /^(?:https?:)?(\/\/[\w.]+\/)/.exec(thumbnailUrl)[1];
+
+// TODO: Make this a utilFunction and use in all Click Through events.
+const trackClickThrough = (e, source, target = "_blank") => {
+  const href = getSourceLink(source);
+
+  const gaEvent = {
+    type: "Click Through",
+    itemId: getItemId(source),
+    title: joinIfArray(getTitle(source)),
+    partner: joinIfArray(getPartner(source)),
+    contributor: joinIfArray(getContributor(source))
+  };
+
+  // e is a React synthetic event
+  e.preventDefault();
+
+  try {
+    // Try tracking a Google Analytics event.
+    trackGaEvent(gaEvent);
+  } catch (error) {
+    // TODO: What is the best way to log an error?
+  } finally {
+    // Open the link, even if the Google Analytics event tracking failed.
+    window.open(href, target);
+  }
+};
 
 const ContentAndMetadata = ({ source }) => {
   const type = source.mainEntity[0]["@type"];
@@ -179,7 +208,8 @@ const ContentAndMetadata = ({ source }) => {
                 <div className={classNames.linkWrapper}>
                   <a
                     href={getSourceLink(source)}
-                    className={`${classNames.sourceLink} clickThrough`}
+                    className={`${classNames.sourceLink}`}
+                    onClick={e => trackClickThrough(e, source)}
                     rel="noopener noreferrer"
                     target="_blank"
                   >
